LayoutTests/fast/viewport/viewport-69-expected.txt

CONSOLE MESSAGE: line 3: Viewport argument key "xinitial-scale" not recognized and ignored.
ALERT: viewport size 100x110 scale 3.2 with limits [3.2, 5] and userScalable true
